Standardizing ConvToConvIntegerAddCastMul
