+++ tests/sparseml/exporters/transforms/test_conv_to_convinteger_add_cast_mul.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 import onnx
 
-from sparseml.exporters.transforms import ConvertQuantizableConvInteger
+from sparseml.exporters.transforms import ConvToConvIntegerAddCastMul
 from tests.sparseml.exporters.transforms.test_onnx_transform import (
     _create_model as _create_model_no_conv,
 )
@@ -101,7 +101,7 @@ def _create_test_model():
 
 def test_convert_quantizable_conv_integer():
     model = _create_test_model()
-    transform = ConvertQuantizableConvInteger()
+    transform = ConvToConvIntegerAddCastMul()
     model = transform(model)
     onnx.checker.check_model(model)
     assert [node.name for node in model.graph.node] == [
@@ -123,7 +123,7 @@ def test_convert_quantizable_conv_integer():
 def test_convert_quantizable_conv_integer_no_conv():
     model_in = _create_model_no_conv()
     nodes_in = [node.name for node in model_in.graph.node]
-    transform = ConvertQuantizableConvInteger()
+    transform = ConvToConvIntegerAddCastMul()
     model_out = transform(model_in)
     onnx.checker.check_model(model_out)
     assert [node.name for node in model_out.graph.node] == nodes_in
